fix(mlx): align sampler / denoiser / CLIP dtypes with WebUI reference
Three dtype mismatches vs the reference (ldm_patched/modules/model_base.py):

1. MLXCFGDenoiser._calc_denoised returned bfloat16
   Reference (model_base.py:194):
     model_output = diffusion_model(...).float()   ← explicit f32 cast
   The sampler loop received bfloat16 denoised, so (x − D)/σ was limited
   to bf16 precision (~0.78%) even though arithmetic was upcast to f32.
   Fix: remove final .astype(mx.bfloat16) → returns float32.

2. Sampler x stored in bfloat16 between steps
   Reference k-diffusion samplers keep x as float32 throughout; only the
   UNet input (xc) is cast to model dtype inside calculate_input.
   Storing x in bf16 introduced ~0.78% quantisation per step — over 20
   steps that compounds to ~17% accumulated error in the latent state.
   Fix: initial x in make_mlx_func is now float32; all .astype(bfloat16)
   casts removed from ODE/SDE x updates in all 7 samplers.  The UNet
   still receives bf16 input via _calc_input's final .astype(bfloat16).

3. CLIP text_projection matmul in bfloat16
   Reference (forge_clip.py:122):
     pooled_output = pooled_output.float() @ text_projection.float()
   The SDXL 1280-dim pooled conditioning vector was projected in bf16,
   degrading the accuracy of the text guidance embedding.
   Fix: cast raw_pooled and tp_w to float32 before the matmul.

Side-effect: all sampler loop bodies are now cleaner — since x and
denoised are both float32, the explicit .astype(mx.float32) guard
casts are no longer needed and have been removed.

+++ mlx_pipeline/clip.py
@@ -251,10 +251,11 @@ class MLXCLIPTextEncoder:
             # np.array() implicitly evaluates — no explicit mx.eval() needed.
             x_np       = np.array(x_normed.astype(mx.float32))         # [B, T, d]
             raw_pooled = mx.array(x_np[batch_idx, eos_pos])            # [B, d]
-            # tp_w is stored pre-transposed to [in, out] so we multiply directly:
-            # pooled = eos_hidden @ tp_w  (matches SDClipModel convention:
-            # clip_g.text_projection is an nn.Parameter used as x @ param)
-            pooled = (raw_pooled.astype(mx.bfloat16) @ self.tp_w)      # [B, d]
+            # Reference (forge_clip.py line 122):
+            #   pooled_output = pooled_output.float() @ text_projection.float()
+            # Project in float32 to match; self.tp_w is [d_in, d_out].
+            pooled = (raw_pooled.astype(mx.float32)
+                      @ self.tp_w.astype(mx.float32))                   # [B, d] f32
 
         # Convert intermediate and pooled back to torch.
         # np.array() implicitly evaluates the MLX graph — no explicit eval needed.
